Initial check-in of GPL'd code from TH3DStudio for "Unified Beta1".
This code is being checked in as-received from TH3Dstudio and is being redistributed, as per the allowances of the GPL. All apparent copyrights are denoted by their owners OR should assumed to be TH3DStudio if no attribuions exist. I claim no copyright and have made no changes to this code, other than redistribution as allowed by the GNU GPL License.